// apps/web/src/http/uazapi/types.ts
export type MediaType =
  | "image"
  | "video"
  | "document"
  | "audio"
  | "myaudio"
  | "ptt"
  | "ptv"
  | "sticker";